+++ features/variety.py
